Move getCloudUrlFromRegion to shared/utils/urls
