+++ apps/code/src/shared/constants/oauth.ts
@@ -19,18 +19,6 @@ export const REGION_LABELS: Record<CloudRegion, string> = {
 export const TOKEN_REFRESH_BUFFER_MS = 30 * 60 * 1000; // 30 minutes before expiry
 export const TOKEN_REFRESH_FORCE_MS = 60 * 1000; // Force refresh when <1 min to expiry, even with active sessions
 
-export function getCloudUrlFromRegion(region: CloudRegion): string {
-  switch (region) {
-    case "us":
-      return "https://us.posthog.com";
-    case "eu":
-      return "https://eu.posthog.com";
-    case "dev":
-      return "http://localhost:8010";
-  }
-}
-
-
 export function getOauthClientIdFromRegion(region: CloudRegion): string {
   switch (region) {
     case "us":
