+++ apps/code/src/shared/utils/urls.ts
@@ -1,5 +1,16 @@
 import { useAuthStore } from "@features/auth/stores/authStore";
-import { getCloudUrlFromRegion } from "@shared/constants/oauth";
+import type { CloudRegion } from "@shared/types/oauth";
+
+export function getCloudUrlFromRegion(region: CloudRegion): string {
+  switch (region) {
+    case "us":
+      return "https://us.posthog.com";
+    case "eu":
+      return "https://eu.posthog.com";
+    case "dev":
+      return "http://localhost:8010";
+  }
+}
 
 export function getPostHogUrl(path: string): string {
   const region = useAuthStore.getState().cloudRegion;
